fix: Replace old Data Import (v12) (#10865)
