--- frappe/core/doctype/data_import/README.md
@@ -1,1 +1,0 @@
-Bulk import / update of data via file upload in Excel or CSV.
