+++ frappe/core/doctype/data_import/data_import.js
@@ -1,324 +1,513 @@
-// Copyright (c) 2017, Frappe Technologies and contributors
+// Copyright (c) 2019, Frappe Technologies and contributors
 // For license information, please see license.txt
 
 frappe.ui.form.on('Data Import', {
-	onload: function(frm) {
-		if (frm.doc.__islocal) {
-			frm.set_value("action", "");
-		}
-
-		frappe.call({
-			method: "frappe.core.doctype.data_import.data_import.get_importable_doctypes",
-			callback: function (r) {
-				let importable_doctypes = r.message;
-				frm.set_query("reference_doctype", function () {
-					return {
-						"filters": {
-							"issingle": 0,
-							"istable": 0,
-							"name": ['in', importable_doctypes]
-						}
-					};
-				});
+	setup(frm) {
+		frappe.realtime.on('data_import_refresh', ({ data_import }) => {
+			frm.import_in_progress = false;
+			if (data_import !== frm.doc.name) return;
+			frappe.model.clear_doc('Data Import', frm.doc.name);
+			frappe.model.with_doc('Data Import', frm.doc.name).then(() => {
+				frm.refresh();
+			});
+		});
+		frappe.realtime.on('data_import_progress', data => {
+			frm.import_in_progress = true;
+			if (data.data_import !== frm.doc.name) {
+				return;
 			}
-		}),
+			let percent = Math.floor((data.current * 100) / data.total);
+			let seconds = Math.floor(data.eta);
+			let minutes = Math.floor(data.eta / 60);
+			let eta_message =
+				// prettier-ignore
+				seconds < 60
+					? __('About {0} seconds remaining', [seconds])
+					: minutes === 1
+						? __('About {0} minute remaining', [minutes])
+						: __('About {0} minutes remaining', [minutes]);
 
-		// should never check public
-		frm.fields_dict["import_file"].df.is_private = 1;
+			let message;
+			if (data.success) {
+				let message_args = [data.current, data.total, eta_message];
+				message =
+					frm.doc.import_type === 'Insert New Records'
+						? __('Importing {0} of {1}, {2}', message_args)
+						: __('Updating {0} of {1}, {2}', message_args);
+			}
+			if (data.skipping) {
+				message = __('Skipping {0} of {1}, {2}', [
+					data.current,
+					data.total,
+					eta_message
+				]);
+			}
+			frm.dashboard.show_progress(__('Import Progress'), percent, message);
+			frm.page.set_indicator(__('In Progress'), 'orange');
 
-		frappe.realtime.on("data_import_progress", function(data) {
-			if (data.data_import === frm.doc.name) {
-				if (data.reload && data.reload === true) {
-					frm.reload_doc();
-				}
-				if (data.progress) {
-					let progress_bar = $(frm.dashboard.progress_area).find(".progress-bar");
-					if (progress_bar) {
-						$(progress_bar).removeClass("progress-bar-danger").addClass("progress-bar-success progress-bar-striped");
-						$(progress_bar).css("width", data.progress + "%");
-					}
-				}
+			// hide progress when complete
+			if (data.current === data.total) {
+				setTimeout(() => {
+					frm.dashboard.hide();
+					frm.refresh();
+				}, 2000);
 			}
 		});
+
+		frm.set_query('reference_doctype', () => {
+			return {
+				filters: {
+					name: ['in', frappe.boot.user.can_import]
+				}
+			};
+		});
+
+		frm.get_field('import_file').df.options = {
+			restrictions: {
+				allowed_file_types: ['.csv', '.xls', '.xlsx']
+			}
+		};
+
+		frm.has_import_file = () => {
+			return frm.doc.import_file || frm.doc.google_sheets_url;
+		};
 	},
 
-	reference_doctype: function(frm){
-		if (frm.doc.reference_doctype) {
-			frappe.model.with_doctype(frm.doc.reference_doctype);
+	refresh(frm) {
+		frm.page.hide_icon_group();
+		frm.trigger('update_indicators');
+		frm.trigger('import_file');
+		frm.trigger('show_import_log');
+		frm.trigger('show_import_warnings');
+		frm.trigger('toggle_submit_after_import');
+		frm.trigger('show_import_status');
+		frm.trigger('show_report_error_button');
+
+		if (frm.doc.status === 'Partial Success') {
+			frm.add_custom_button(__('Export Errored Rows'), () =>
+				frm.trigger('export_errored_rows')
+			);
+		}
+
+		if (frm.doc.status.includes('Success')) {
+			frm.add_custom_button(
+				__('Go to {0} List', [frm.doc.reference_doctype]),
+				() => frappe.set_route('List', frm.doc.reference_doctype)
+			);
 		}
 	},
 
-	refresh: function(frm) {
+	onload_post_render(frm) {
+		frm.trigger('update_primary_action');
+	},
+
+	update_primary_action(frm) {
+		if (frm.is_dirty()) {
+			frm.enable_save();
+			return;
+		}
 		frm.disable_save();
-		frm.dashboard.clear_headline();
-		if (frm.doc.reference_doctype && !frm.doc.import_file) {
-			frm.page.set_indicator(__('Attach file'), 'orange');
-		} else {
-			if (frm.doc.import_status) {
-				const listview_settings = frappe.listview_settings['Data Import'];
-				const indicator = listview_settings.get_indicator(frm.doc);
-
-				frm.page.set_indicator(indicator[0], indicator[1]);
-
-				if (frm.doc.import_status === "In Progress") {
-					frm.dashboard.add_progress("Data Import Progress", "0");
-					frm.set_read_only();
-					frm.refresh_fields();
-				}
+		if (frm.doc.status !== 'Success') {
+			if (!frm.is_new() && (frm.has_import_file())) {
+				let label =
+					frm.doc.status === 'Pending' ? __('Start Import') : __('Retry');
+				frm.page.set_primary_action(label, () => frm.events.start_import(frm));
+			} else {
+				frm.page.set_primary_action(__('Save'), () => frm.save());
 			}
 		}
+	},
 
-		if (frm.doc.reference_doctype) {
-			frappe.model.with_doctype(frm.doc.reference_doctype);
+	update_indicators(frm) {
+		const indicator = frappe.get_indicator(frm.doc);
+		if (indicator) {
+			frm.page.set_indicator(indicator[0], indicator[1]);
+		} else {
+			frm.page.clear_indicator();
 		}
+	},
 
-		if(frm.doc.action == "Insert new records" || frm.doc.action == "Update records") {
-			frm.set_df_property("action", "read_only", 1);
+	show_import_status(frm) {
+		let import_log = JSON.parse(frm.doc.import_log || '[]');
+		let successful_records = import_log.filter(log => log.success);
+		let failed_records = import_log.filter(log => !log.success);
+		if (successful_records.length === 0) return;
+
+		let message;
+		if (failed_records.length === 0) {
+			let message_args = [successful_records.length];
+			if (frm.doc.import_type === 'Insert New Records') {
+				message =
+					successful_records.length > 1
+						? __('Successfully imported {0} records.', message_args)
+						: __('Successfully imported {0} record.', message_args);
+			} else {
+				message =
+					successful_records.length > 1
+						? __('Successfully updated {0} records.', message_args)
+						: __('Successfully updated {0} record.', message_args);
+			}
+		} else {
+			let message_args = [successful_records.length, import_log.length];
+			if (frm.doc.import_type === 'Insert New Records') {
+				message =
+					successful_records.length > 1
+						? __('Successfully imported {0} records out of {1}. Click on Export Errored Rows, fix the errors and import again.', message_args)
+						: __('Successfully imported {0} record out of {1}. Click on Export Errored Rows, fix the errors and import again.', message_args);
+			} else {
+				message =
+					successful_records.length > 1
+						? __('Successfully updated {0} records out of {1}. Click on Export Errored Rows, fix the errors and import again.', message_args)
+						: __('Successfully updated {0} record out of {1}. Click on Export Errored Rows, fix the errors and import again.', message_args);
+			}
 		}
+		frm.dashboard.set_headline(message);
+	},
 
-		frm.add_custom_button(__("Help"), function() {
-			frappe.help.show_video("6wiriRKPhmg");
+	show_report_error_button(frm) {
+		if (frm.doc.status === 'Error') {
+			frappe.db
+				.get_list('Error Log', {
+					filters: { method: frm.doc.name },
+					fields: ['method', 'error'],
+					order_by: 'creation desc',
+					limit: 1
+				})
+				.then(result => {
+					if (result.length > 0) {
+						frm.add_custom_button('Report Error', () => {
+							let fake_xhr = {
+								responseText: JSON.stringify({
+									exc: result[0].error
+								})
+							};
+							frappe.request.report_error(fake_xhr, {});
+						});
+					}
+				});
+		}
+	},
+
+	start_import(frm) {
+		frm
+			.call({
+				method: 'form_start_import',
+				args: { data_import: frm.doc.name },
+				btn: frm.page.btn_primary
+			})
+			.then(r => {
+				if (r.message === true) {
+					frm.disable_save();
+				}
+			});
+	},
+
+	download_template(frm) {
+		frappe.require('/assets/js/data_import_tools.min.js', () => {
+			frm.data_exporter = new frappe.data_import.DataExporter(
+				frm.doc.reference_doctype,
+				frm.doc.import_type
+			);
 		});
+	},
 
-		if (frm.doc.reference_doctype && frm.doc.docstatus === 0) {
-			frm.add_custom_button(__("Download template"), function() {
-				frappe.data_import.download_dialog(frm).show();
+	reference_doctype(frm) {
+		frm.trigger('toggle_submit_after_import');
+	},
+
+	toggle_submit_after_import(frm) {
+		frm.toggle_display('submit_after_import', false);
+		let doctype = frm.doc.reference_doctype;
+		if (doctype) {
+			frappe.model.with_doctype(doctype, () => {
+				let meta = frappe.get_meta(doctype);
+				frm.toggle_display('submit_after_import', meta.is_submittable);
 			});
 		}
+	},
 
-		if (frm.doc.reference_doctype && frm.doc.import_file && frm.doc.total_rows &&
-			frm.doc.docstatus === 0 && (!frm.doc.import_status || frm.doc.import_status == "Failed")) {
-			frm.page.set_primary_action(__("Start Import"), function() {
-				frappe.call({
-					btn: frm.page.btn_primary,
-					method: "frappe.core.doctype.data_import.data_import.import_data",
-					args: {
-						data_import: frm.doc.name
-					}
-				});
-			}).addClass('btn btn-primary');
-		}
-
-		if (frm.doc.log_details) {
-			frm.events.create_log_table(frm);
+	google_sheets_url(frm) {
+		if (!frm.is_dirty()) {
+			frm.trigger('import_file');
 		} else {
-			$(frm.fields_dict.import_log.wrapper).empty();
+			frm.trigger('update_primary_action');
 		}
 	},
 
-	action: function(frm) {
-		if(!frm.doc.action) return;
-		if(!frm.doc.reference_doctype) {
-			frappe.msgprint(__("Please select document type first."));
-			frm.set_value("action", "");
+	refresh_google_sheet(frm) {
+		frm.trigger('import_file');
+	},
+
+	import_file(frm) {
+		frm.toggle_display('section_import_preview', frm.has_import_file());
+		if (!frm.has_import_file()) {
+			frm.get_field('import_preview').$wrapper.empty();
+			return;
+		} else {
+			frm.trigger('update_primary_action');
+		}
+
+		// load import preview
+		frm.get_field('import_preview').$wrapper.empty();
+		$('<span class="text-muted">')
+			.html(__('Loading import file...'))
+			.appendTo(frm.get_field('import_preview').$wrapper);
+
+		frm
+			.call({
+				method: 'get_preview_from_template',
+				args: {
+					data_import: frm.doc.name,
+					import_file: frm.doc.import_file,
+					google_sheets_url: frm.doc.google_sheets_url
+				},
+				error_handlers: {
+					TimestampMismatchError() {
+						// ignore this error
+					}
+				}
+			})
+			.then(r => {
+				let preview_data = r.message;
+				frm.events.show_import_preview(frm, preview_data);
+				frm.events.show_import_warnings(frm, preview_data);
+			});
+	},
+
+	show_import_preview(frm, preview_data) {
+		let import_log = JSON.parse(frm.doc.import_log || '[]');
+
+		if (
+			frm.import_preview &&
+			frm.import_preview.doctype === frm.doc.reference_doctype
+		) {
+			frm.import_preview.preview_data = preview_data;
+			frm.import_preview.import_log = import_log;
+			frm.import_preview.refresh();
 			return;
 		}
 
-		if(frm.doc.action == "Insert new records") {
-			frm.doc.insert_new = 1;
-		} else if (frm.doc.action == "Update records"){
-			frm.doc.overwrite = 1;
+		frappe.require('/assets/js/data_import_tools.min.js', () => {
+			frm.import_preview = new frappe.data_import.ImportPreview({
+				wrapper: frm.get_field('import_preview').$wrapper,
+				doctype: frm.doc.reference_doctype,
+				preview_data,
+				import_log,
+				frm,
+				events: {
+					remap_column(changed_map) {
+						let template_options = JSON.parse(frm.doc.template_options || '{}');
+						template_options.column_to_field_map = template_options.column_to_field_map || {};
+						Object.assign(template_options.column_to_field_map, changed_map);
+						frm.set_value('template_options', JSON.stringify(template_options));
+						frm.save().then(() => frm.trigger('import_file'));
+					}
+				}
+			});
+		});
+	},
+
+	export_errored_rows(frm) {
+		open_url_post(
+			'/api/method/frappe.core.doctype.data_import.data_import.download_errored_template',
+			{
+				data_import_name: frm.doc.name
+			}
+		);
+	},
+
+	show_import_warnings(frm, preview_data) {
+		let warnings = JSON.parse(frm.doc.template_warnings || '[]');
+		warnings = warnings.concat(preview_data.warnings || []);
+
+		frm.toggle_display('import_warnings_section', warnings.length > 0);
+		if (warnings.length === 0) {
+			frm.get_field('import_warnings').$wrapper.html('');
+			return;
 		}
-		frm.save();
+
+		// group warnings by row
+		let warnings_by_row = {};
+		let other_warnings = [];
+		for (let warning of warnings) {
+			if (warning.row) {
+				warnings_by_row[warning.row] = warnings_by_row[warning.row] || [];
+				warnings_by_row[warning.row].push(warning);
+			} else {
+				other_warnings.push(warning);
+			}
+		}
+
+		let html = '';
+		html += Object.keys(warnings_by_row)
+			.map(row_number => {
+				let message = warnings_by_row[row_number]
+					.map(w => {
+						if (w.field) {
+							return `<li>${w.field.label}: ${w.message}</li>`;
+						}
+						return `<li>${w.message}</li>`;
+					})
+					.join('');
+				return `
+				<div class="warning" data-row="${row_number}">
+					<h5 class="text-uppercase">${__('Row {0}', [row_number])}</h5>
+					<div class="body"><ul>${message}</ul></div>
+				</div>
+			`;
+			})
+			.join('');
+
+		html += other_warnings
+			.map(warning => {
+				let header = '';
+				if (warning.col) {
+					header = __('Column {0}', [warning.col]);
+				}
+				return `
+					<div class="warning" data-col="${warning.col}">
+						<h5 class="text-uppercase">${header}</h5>
+						<div class="body">${warning.message}</div>
+					</div>
+				`;
+			})
+			.join('');
+		frm.get_field('import_warnings').$wrapper.html(`
+			<div class="row">
+				<div class="col-sm-10 warnings">${html}</div>
+			</div>
+		`);
 	},
 
-	only_update: function(frm) {
-		frm.save();
+	show_failed_logs(frm) {
+		frm.trigger('show_import_log');
 	},
 
-	submit_after_import: function(frm) {
-		frm.save();
+	show_import_log(frm) {
+		let import_log = JSON.parse(frm.doc.import_log || '[]');
+		let logs = import_log;
+		frm.toggle_display('import_log', false);
+		frm.toggle_display('import_log_section', logs.length > 0);
+
+		if (logs.length === 0) {
+			frm.get_field('import_log_preview').$wrapper.empty();
+			return;
+		}
+
+		let rows = logs
+			.map(log => {
+				let html = '';
+				if (log.success) {
+					if (frm.doc.import_type === 'Insert New Records') {
+						html = __('Successfully imported {0}', [
+							`<span class="underline">${frappe.utils.get_form_link(
+								frm.doc.reference_doctype,
+								log.docname,
+								true
+							)}<span>`
+						]);
+					} else {
+						html = __('Successfully updated {0}', [
+							`<span class="underline">${frappe.utils.get_form_link(
+								frm.doc.reference_doctype,
+								log.docname,
+								true
+							)}<span>`
+						]);
+					}
+				} else {
+					let messages = log.messages
+						.map(JSON.parse)
+						.map(m => {
+							let title = m.title ? `<strong>${m.title}</strong>` : '';
+							let message = m.message ? `<div>${m.message}</div>` : '';
+							return title + message;
+						})
+						.join('');
+					let id = frappe.dom.get_unique_id();
+					html = `${messages}
+						<button class="btn btn-default btn-xs" type="button" data-toggle="collapse" data-target="#${id}" aria-expanded="false" aria-controls="${id}" style="margin-top: 15px;">
+							${__('Show Traceback')}
+						</button>
+						<div class="collapse" id="${id}" style="margin-top: 15px;">
+							<div class="well">
+								<pre>${log.exception}</pre>
+							</div>
+						</div>`;
+				}
+				let indicator_color = log.success ? 'green' : 'red';
+				let title = log.success ? __('Success') : __('Failure');
+
+				if (frm.doc.show_failed_logs && log.success) {
+					return '';
+				}
+
+				return `<tr>
+					<td>${log.row_indexes.join(', ')}</td>
+					<td>
+						<div class="indicator ${indicator_color}">${title}</div>
+					</td>
+					<td>
+						${html}
+					</td>
+				</tr>`;
+			})
+			.join('');
+
+		if (!rows && frm.doc.show_failed_logs) {
+			rows = `<tr><td class="text-center text-muted" colspan=3>
+				${__('No failed logs')}
+			</td></tr>`;
+		}
+
+		frm.get_field('import_log_preview').$wrapper.html(`
+			<table class="table table-bordered">
+				<tr class="text-muted">
+					<th width="10%">${__('Row Number')}</th>
+					<th width="10%">${__('Status')}</th>
+					<th width="80%">${__('Message')}</th>
+				</tr>
+				${rows}
+			</table>
+		`);
 	},
 
-	skip_errors: function(frm) {
-		frm.save();
-	},
+	show_missing_link_values(frm, missing_link_values) {
+		let can_be_created_automatically = missing_link_values.every(
+			d => d.has_one_mandatory_field
+		);
 
-	ignore_encoding_errors: function(frm) {
-		frm.save();
-	},
+		let html = missing_link_values
+			.map(d => {
+				let doctype = d.doctype;
+				let values = d.missing_values;
+				return `
+					<h5>${doctype}</h5>
+					<ul>${values.map(v => `<li>${v}</li>`).join('')}</ul>
+				`;
+			})
+			.join('');
 
-	no_email: function(frm) {
-		frm.save();
-	},
-
-	show_only_errors: function(frm) {
-		frm.events.create_log_table(frm);
-	},
-
-	create_log_table: function(frm) {
-		let msg = JSON.parse(frm.doc.log_details);
-		var $log_wrapper = $(frm.fields_dict.import_log.wrapper).empty();
-		$(frappe.render_template("log_details", {
-			data: msg.messages,
-			import_status: frm.doc.import_status,
-			show_only_errors: frm.doc.show_only_errors,
-		})).appendTo($log_wrapper);
+		if (can_be_created_automatically) {
+			// prettier-ignore
+			let message = __('There are some linked records which needs to be created before we can import your file. Do you want to create the following missing records automatically?');
+			frappe.confirm(message + html, () => {
+				frm
+					.call('create_missing_link_values', {
+						missing_link_values
+					})
+					.then(r => {
+						let records = r.message;
+						frappe.msgprint(
+							__('Created {0} records successfully.', [records.length])
+						);
+					});
+			});
+		} else {
+			frappe.msgprint(
+				// prettier-ignore
+				__('The following records needs to be created before we can import your file.') + html
+			);
+		}
 	}
 });
-
-frappe.provide('frappe.data_import');
-frappe.data_import.download_dialog = function(frm) {
-	var dialog;
-	const filter_fields = df => frappe.model.is_value_type(df) && !df.hidden;
-	const get_fields = dt => frappe.meta.get_docfields(dt).filter(filter_fields);
-
-	const get_doctype_checkbox_fields = () => {
-		return dialog.fields.filter(df => df.fieldname.endsWith('_fields'))
-			.map(df => dialog.fields_dict[df.fieldname]);
-	};
-
-	const doctype_fields = get_fields(frm.doc.reference_doctype)
-		.map(df => {
-			let reqd = (df.reqd || df.fieldname == 'naming_series') ? 1 : 0;
-			return {
-				label: df.label,
-				reqd: reqd,
-				danger: reqd,
-				value: df.fieldname,
-				checked: 1
-			};
-		});
-
-	let fields = [
-		{
-			"label": __("Select Columns"),
-			"fieldname": "select_columns",
-			"fieldtype": "Select",
-			"options": "All\nMandatory\nManually",
-			"reqd": 1,
-			"onchange": function() {
-				const fields = get_doctype_checkbox_fields();
-				fields.map(f => f.toggle(true));
-				if(this.value == 'Mandatory' || this.value == 'Manually') {
-					checkbox_toggle(true);
-					fields.map(multicheck_field => {
-						multicheck_field.options.map(option => {
-							if(!option.reqd) return;
-							$(multicheck_field.$wrapper).find(`:checkbox[data-unit="${option.value}"]`)
-								.prop('checked', false)
-								.trigger('click');
-						});
-					});
-				} else if(this.value == 'All'){
-					$(dialog.body).find(`[data-fieldtype="MultiCheck"] :checkbox`)
-						.prop('disabled', true);
-				}
-			}
-		},
-		{
-			"label": __("File Type"),
-			"fieldname": "file_type",
-			"fieldtype": "Select",
-			"options": "Excel\nCSV",
-			"default": "Excel"
-		},
-		{
-			"label": __("Download with Data"),
-			"fieldname": "with_data",
-			"fieldtype": "Check",
-			"hidden": !frm.doc.overwrite,
-			"default": 1
-		},
-		{
-			"label": __("Select All"),
-			"fieldname": "select_all",
-			"fieldtype": "Button",
-			"depends_on": "eval:doc.select_columns=='Manually'",
-			click: function() {
-				checkbox_toggle();
-			}
-		},
-		{
-			"label": __("Unselect All"),
-			"fieldname": "unselect_all",
-			"fieldtype": "Button",
-			"depends_on": "eval:doc.select_columns=='Manually'",
-			click: function() {
-				checkbox_toggle(true);
-			}
-		},
-		{
-			"label": frm.doc.reference_doctype,
-			"fieldname": "doctype_fields",
-			"fieldtype": "MultiCheck",
-			"options": doctype_fields,
-			"columns": 2,
-			"hidden": 1
-		}
-	];
-
-	const child_table_fields = frappe.meta.get_table_fields(frm.doc.reference_doctype)
-		.map(df => {
-			return {
-				"label": df.options,
-				"fieldname": df.fieldname + '_fields',
-				"fieldtype": "MultiCheck",
-				"options": frappe.meta.get_docfields(df.options)
-					.filter(filter_fields)
-					.map(df => ({
-						label: df.label,
-						reqd: df.reqd ? 1 : 0,
-						value: df.fieldname,
-						checked: 1,
-						danger: df.reqd
-					})),
-				"columns": 2,
-				"hidden": 1
-			};
-		});
-
-	fields = fields.concat(child_table_fields);
-
-	dialog = new frappe.ui.Dialog({
-		title: __('Download Template'),
-		fields: fields,
-		primary_action: function(values) {
-			var data = values;
-			if (frm.doc.reference_doctype) {
-				var export_params = () => {
-					let columns = {};
-					if(values.select_columns) {
-						columns = get_doctype_checkbox_fields().reduce((columns, field) => {
-							const options = field.get_checked_options();
-							columns[field.df.label] = options;
-							return columns;
-						}, {});
-					}
-
-					return {
-						doctype: frm.doc.reference_doctype,
-						parent_doctype: frm.doc.reference_doctype,
-						select_columns: JSON.stringify(columns),
-						with_data: frm.doc.overwrite && data.with_data,
-						all_doctypes: true,
-						file_type: data.file_type,
-						template: true
-					};
-				};
-				let get_template_url = '/api/method/frappe.core.doctype.data_export.exporter.export_data';
-				open_url_post(get_template_url, export_params());
-			} else {
-				frappe.msgprint(__("Please select the Document Type."));
-			}
-			dialog.hide();
-		},
-		primary_action_label: __('Download')
-	});
-
-	$(dialog.body).find('div[data-fieldname="select_all"], div[data-fieldname="unselect_all"]')
-		.wrapAll('<div class="inline-buttons" />');
-	const button_container = $(dialog.body).find('.inline-buttons');
-	button_container.addClass('flex');
-	$(button_container).find('.frappe-control').map((index, button) => {
-		$(button).css({"margin-right": "1em"});
-	});
-
-	function checkbox_toggle(checked=false) {
-		$(dialog.body).find('[data-fieldtype="MultiCheck"]').map((index, element) => {
-			$(element).find(`:checkbox`).prop("checked", checked).trigger('click');
-		});
-	}
-
-	return dialog;
-};
