+++ frappe/core/doctype/data_import/fixtures/sample_import_file_without_mandatory.csv
