+++ frappe/core/doctype/data_import/test_exporter.py
@@ -6,8 +6,8 @@ from __future__ import unicode_literals
 import unittest
 import six
 import frappe
-from frappe.core.doctype.data_import_beta.exporter import Exporter
-from frappe.core.doctype.data_import_beta.test_importer import (
+from frappe.core.doctype.data_import.exporter import Exporter
+from frappe.core.doctype.data_import.test_importer import (
 	create_doctype_if_not_exists,
 )
