+++ frappe/core/doctype/data_import_legacy/data_import_legacy.json
@@ -1,0 +1,218 @@
+{
+ "actions": [],
+ "allow_copy": 1,
+ "creation": "2020-06-11 16:13:23.813709",
+ "doctype": "DocType",
+ "document_type": "Document",
+ "editable_grid": 1,
+ "engine": "InnoDB",
+ "field_order": [
+  "reference_doctype",
+  "action",
+  "insert_new",
+  "overwrite",
+  "only_update",
+  "section_break_4",
+  "import_file",
+  "column_break_4",
+  "error_file",
+  "section_break_6",
+  "skip_errors",
+  "submit_after_import",
+  "ignore_encoding_errors",
+  "no_email",
+  "import_detail",
+  "import_status",
+  "show_only_errors",
+  "import_log",
+  "log_details",
+  "amended_from",
+  "total_rows",
+  "amended_from"
+ ],
+ "fields": [
+  {
+   "fieldname": "reference_doctype",
+   "fieldtype": "Link",
+   "ignore_user_permissions": 1,
+   "in_list_view": 1,
+   "label": "Document Type",
+   "options": "DocType",
+   "reqd": 1
+  },
+  {
+   "fieldname": "action",
+   "fieldtype": "Select",
+   "label": "Action",
+   "options": "Insert new records\nUpdate records",
+   "reqd": 1
+  },
+  {
+   "default": "0",
+   "depends_on": "eval:!doc.overwrite",
+   "description": "New data will be inserted.",
+   "fieldname": "insert_new",
+   "fieldtype": "Check",
+   "hidden": 1,
+   "label": "Insert new records",
+   "set_only_once": 1
+  },
+  {
+   "default": "0",
+   "depends_on": "eval:!doc.insert_new",
+   "description": "If you are updating/overwriting already created records.",
+   "fieldname": "overwrite",
+   "fieldtype": "Check",
+   "hidden": 1,
+   "label": "Update records",
+   "set_only_once": 1
+  },
+  {
+   "default": "0",
+   "depends_on": "overwrite",
+   "description": "If you don't want to create any new records while updating the older records.",
+   "fieldname": "only_update",
+   "fieldtype": "Check",
+   "label": "Don't create new records"
+  },
+  {
+   "depends_on": "eval:(!doc.__islocal)",
+   "fieldname": "section_break_4",
+   "fieldtype": "Section Break"
+  },
+  {
+   "fieldname": "import_file",
+   "fieldtype": "Attach",
+   "label": "Attach file for Import"
+  },
+  {
+   "fieldname": "column_break_4",
+   "fieldtype": "Column Break"
+  },
+  {
+   "depends_on": "eval: doc.import_status == \"Partially Successful\"",
+   "description": "This is the template file generated with only the rows having some error. You should use this file for correction and import.",
+   "fieldname": "error_file",
+   "fieldtype": "Attach",
+   "label": "Generated File"
+  },
+  {
+   "depends_on": "eval:(!doc.__islocal)",
+   "fieldname": "section_break_6",
+   "fieldtype": "Section Break"
+  },
+  {
+   "default": "0",
+   "description": "If this is checked, rows with valid data will be imported and invalid rows will be dumped into a new file for you to import later.",
+   "fieldname": "skip_errors",
+   "fieldtype": "Check",
+   "label": "Skip rows with errors"
+  },
+  {
+   "default": "0",
+   "fieldname": "submit_after_import",
+   "fieldtype": "Check",
+   "label": "Submit after importing"
+  },
+  {
+   "default": "0",
+   "fieldname": "ignore_encoding_errors",
+   "fieldtype": "Check",
+   "label": "Ignore encoding errors"
+  },
+  {
+   "default": "1",
+   "fieldname": "no_email",
+   "fieldtype": "Check",
+   "label": "Do not send Emails"
+  },
+  {
+   "collapsible": 1,
+   "collapsible_depends_on": "eval: doc.import_status == \"Failed\"",
+   "depends_on": "import_status",
+   "fieldname": "import_detail",
+   "fieldtype": "Section Break",
+   "label": "Import Log"
+  },
+  {
+   "fieldname": "import_status",
+   "fieldtype": "Select",
+   "label": "Import Status",
+   "options": "\nSuccessful\nFailed\nIn Progress\nPartially Successful",
+   "read_only": 1
+  },
+  {
+   "allow_on_submit": 1,
+   "default": "1",
+   "fieldname": "show_only_errors",
+   "fieldtype": "Check",
+   "label": "Show only errors",
+   "no_copy": 1,
+   "print_hide": 1
+  },
+  {
+   "allow_on_submit": 1,
+   "depends_on": "import_status",
+   "fieldname": "import_log",
+   "fieldtype": "HTML",
+   "label": "Import Log"
+  },
+  {
+   "allow_on_submit": 1,
+   "fieldname": "log_details",
+   "fieldtype": "Code",
+   "hidden": 1,
+   "label": "Log Details",
+   "read_only": 1
+  },
+  {
+   "fieldname": "amended_from",
+   "fieldtype": "Link",
+   "label": "Amended From",
+   "no_copy": 1,
+   "options": "Data Import",
+   "print_hide": 1,
+   "read_only": 1
+  },
+  {
+   "fieldname": "total_rows",
+   "fieldtype": "Int",
+   "hidden": 1,
+   "label": "Total Rows",
+   "read_only": 1
+  },
+  {
+   "fieldname": "amended_from",
+   "fieldtype": "Link",
+   "label": "Amended From",
+   "no_copy": 1,
+   "options": "Data Import Legacy",
+   "print_hide": 1,
+   "read_only": 1
+  }
+ ],
+ "is_submittable": 1,
+ "links": [],
+ "max_attachments": 1,
+ "modified": "2020-06-11 16:13:23.813709",
+ "modified_by": "Administrator",
+ "module": "Core",
+ "name": "Data Import Legacy",
+ "owner": "Administrator",
+ "permissions": [
+  {
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "read": 1,
+   "role": "System Manager",
+   "share": 1,
+   "submit": 1,
+   "write": 1
+  }
+ ],
+ "sort_field": "modified",
+ "sort_order": "DESC",
+ "track_changes": 1,
+ "track_seen": 1
+}
