+++ frappe/patches.txt
@@ -268,3 +268,4 @@ execute:frappe.delete_doc_if_exists('DocType', 'GCalendar Settings')
 frappe.patches.v12_0.remove_parent_and_parenttype_from_print_formats
 execute:from frappe.desk.page.setup_wizard.install_fixtures import update_genders;update_genders()
 frappe.patches.v12_0.email_unsubscribe
+frappe.patches.v12_0.replace_old_data_import
