+++ frappe/public/js/frappe/db.js
@@ -91,12 +91,26 @@ frappe.db = {
 		});
 	},
 	count: function(doctype, args={}) {
-		return new Promise(resolve => {
-			frappe.call({
-				method: 'frappe.client.get_count',
-				type: 'GET',
-				args: Object.assign(args, { doctype })
-			}).then(r => resolve(r.message));
+		let filters = args.filters || {};
+		const with_child_table_filter = Array.isArray(filters) && filters.some(filter => {
+			return filter[0] !== doctype;
+		});
+
+		const fields = [
+			// cannot break this line as it adds extra \n's and \t's which breaks the query
+			`count(${with_child_table_filter ? 'distinct': ''} ${frappe.model.get_full_column_name('name', doctype)}) AS total_count`
+		];
+
+		return frappe.call({
+			type: 'GET',
+			method: 'frappe.desk.reportview.get',
+			args: {
+				doctype,
+				filters,
+				fields,
+			}
+		}).then(r => {
+			return r.message.values[0][0];
 		});
 	},
 	get_link_options(doctype, txt = '', filters={}) {
